+++ tests/integration/test_init_rapyer.py
@@ -1,0 +1,32 @@
+import pytest
+import redis
+
+import rapyer
+from rapyer.base import REDIS_MODELS
+
+
+@pytest.mark.asyncio
+async def test_init_rapyer_with_url__redis_is_available():
+    # Arrange
+    redis_url = "redis://localhost:6370"
+
+    # Act
+    await rapyer.init_rapyer(redis_url)
+
+    for model in REDIS_MODELS:
+        await model.Meta.redis.ping()
+        assert model.Meta.redis.connection_pool.connection_kwargs["decode_responses"]
+
+
+@pytest.mark.asyncio
+async def test_init_rapyer_with_client__redis_is_available():
+    # Arrange
+    created_client = redis.asyncio.from_url(
+        "redis://localhost:6370/1", decode_responses=True
+    )
+
+    # Act
+    await rapyer.init_rapyer(created_client)
+
+    for model in REDIS_MODELS:
+        await model.Meta.redis.ping()