+++ tests/models/functionality_types.py
@@ -3,6 +3,7 @@ from enum import Enum
 from typing import List
 
 from pydantic import Field
+
 from rapyer.base import AtomicRedisModel